Add files via upload
The objective of this project is to conduct comprehensive analysis of the IMDB MOVIE 
dataset. This analysis focus on serval key aspect of the dataset like Genre, language, directors, Duration, Story 
and Budget etc. our main goal is to find a relation between all these aspect and how movie rating and success 
define on these aspect so that in future other movie directors can use this analysis for movie making and make 
a good movies and get their good rating and success